+++ ChangeLog.md
@@ -3,7 +3,8 @@
 - Online help will be included in the released package as markdown. 
   Html based help is only available online on GitHub
 - UI tooltips improved
-- zoom sliders added
+- Zoom sliders added
+- Zoom with cursor keys at mouse position supported
 
 ## 2.7.0 - 2017-12-26